@@ -1,22 +1,26 @@
 {
   "name": "vue-chat",
-  "version": "0.0.2",
+  "version": "1.0.0",
   "description": "chat example by vue.js + webpack",
   "scripts": {
-    "dev": "webpack --watch",
-    "build": "webpack"
+    "dev": "webpack-dev-server --hot --inline",
+    "build": "webpack -p"
   },
   "repository": {
-  	"type": "git",
-  	"url": "git+https://github.com/Coffcer/vue-chat.git"
+    "type": "git",
+    "url": "git+https://github.com/Coffcer/vue-chat.git"
   },
   "dependencies": {
-    "vue": "^0.12.8"
+    "vue": "^1.0.10"
   },
   "devDependencies": {
     "autoprefixer-loader": "^2.0.0",
-    "babel-core": "^5.8.23",
-    "babel-loader": "^5.3.2",
+    "babel": "^6.3.13",
+    "babel-core": "^6.3.21",
+    "babel-loader": "^6.2.0",
+    "babel-plugin-transform-runtime": "^6.3.13",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-runtime": "^5.8.34",
     "css-loader": "^0.16.0",
     "file-loader": "^0.8.4",
     "html-loader": "^0.3.0",
@@ -24,10 +28,13 @@
     "less-loader": "^2.2.0",
     "style-loader": "^0.12.3",
     "url-loader": "^0.5.6",
-    "webpack": "^1.12.0"
+    "vue-hot-reload-api": "^1.2.2",
+    "vue-loader": "^7.2.0",
+    "webpack": "^1.12.0",
+    "webpack-dev-server": "^1.14.0"
   },
   "keywords": [
-  	"vue",
-  	"webpack"
+    "vue",
+    "webpack"
   ]
 }
