@@ -18,21 +18,12 @@ if (!localStorage.getItem(key)) {
                 id: 2,
                 name: '示例介绍',
                 img: 'build/images/2.png'
-            },
-            {
+            },{
                 id: 3,
-                name: '邓紫棋',
+                name: 'webpack',
                 img: 'build/images/3.jpg'
-            },
-            {
-                id: 4,
-                name: 'Webpack',
-                img: 'build/images/4.jpg'
             }
         ],
-        
-        // 当前会话ID
-        sessionId: 0,
 
         // 会话列表
         sessionList: [
@@ -40,7 +31,7 @@ if (!localStorage.getItem(key)) {
                 userId: 2,
                 messages: [
                     {
-                        text: 'Hello，这是一个基于Vue + Webpack构建的简单chat示例，聊天记录保存在localStorge。简单演示了Vue的component、filter、directive、computed以及组件间的事件通讯，代码非常少。',
+                        text: 'Hello，这是一个基于Vue + Webpack构建的简单chat示例，聊天记录保存在localStorge。简单演示了Vue的基础特性和webpack配置。',
                         date: now
                     }, 
                     {
@@ -51,15 +42,6 @@ if (!localStorage.getItem(key)) {
             },
             {
                 userId: 3,
-                messages: [
-                    {
-                        text: '听说下雨天，Vue和Webpack更配喔',
-                        date: now
-                    }
-                ]
-            },
-            {
-                userId: 4,
                 messages: []
             }
         ],
@@ -68,7 +50,7 @@ if (!localStorage.getItem(key)) {
     localStorage.setItem(key, JSON.stringify(data));
 }
 
-module.exports = {
+export default {
     
     fetch () {
         return JSON.parse(localStorage.getItem(key));
