@@ -2,29 +2,36 @@ var path = require('path');
 
 module.exports = {
     // 入口
-    entry: {
-        script: './src/chat'
-    },
+    entry: './src/main',
     // 输出
     output: {
-        path: path.resolve('./build'),
+        path: path.join(__dirname, './build'),
         filename: '[name].js',
-        publicPath: '/build'
+        publicPath: '/build/'
     },
     module: {
         // 加载器
         loaders: [
-            // 使用Babel转换ES6，排除node_modules目录下的js
+            { test: /\.vue$/, loader: 'vue' },
             { test: /\.js$/, loader: 'babel', exclude: /node_modules/ },
             { test: /\.css$/, loader: 'style!css!autoprefixer'},
             { test: /\.less/, loader: 'style!css!autoprefixer!less'},
             { test: /\.(png|jpg|gif)$/, loader: 'url-loader'},
-            { test: /\.(html|tpl)$/, loader: 'html-loader' }
+            { test: /\.(html|tpl)$/, loader: 'html-loader' },
         ]
+    },
+    vue: {
+        loaders: {
+            css: 'style!css!autoprefixer!less'
+        }
+    },
+    babel: {
+        presets: ['es2015'],
+        plugins: ['transform-runtime']
     },
     resolve: {
         // require时省略的扩展名，如：require('module') 不需要module.js
-        extensions: ['', '.js'],
+        extensions: ['', '.js', '.vue'],
         // 别名
         alias: {
             filter: path.join(__dirname, './src/filters'),
